+++ themes/commerce/components/BlogPostListScroll.js
@@ -4,7 +4,8 @@ import { useEffect, useRef, useState } from 'react'
 import CONFIG from '../config'
 import BlogPostListEmpty from './BlogPostListEmpty'
 import ProductCard from './ProductCard'
-import { getListByPage } from '@/lib/utils/post'
+
+import { getListByPage } from '@/lib/utils'
 
 /**
  * 博客列表滚动分页
