+++ themes/heo/components/BlogPostListScroll.js
@@ -4,7 +4,8 @@ import { useEffect, useRef, useState } from 'react'
 import CONFIG from '../config'
 import BlogPostCard from './BlogPostCard'
 import BlogPostListEmpty from './BlogPostListEmpty'
-import { getListByPage } from '@/lib/utils/post'
+
+import { getListByPage } from '@/lib/utils'
 
 /**
  * 博客列表滚动分页
